order/requirements.txt
django==4.2.7
djangorestframework==3.14.0
gunicorn==20.1.0
requests==2.26.0